```text feat: Add columns 'deletedAt' and 'quantity' to Portfolio table
This commit adds the columns 'deletedAt' and 'quantity' to the Portfolio table in the database. The 'deletedAt' column is of type TIMESTAMP(3) and is used to mark a portfolio as deleted. The 'quantity' column is of type INTEGER and has a default value of 0. These changes enhance the functionality of the Portfolio table and allow for better management of portfolios.